change tool names and structure
