+++ c2pa_manifest_tmpl.example.json
@@ -1,6 +1,6 @@
 {
   "ta_url": "http://timestamp.digicert.com",
-  "claim_generator": "integrity-v2",
+  "claim_generator": "Starling Lab Integrity v2",
   "assertions": [
     {
       "label": "stds.schema-org.CreativeWork",
