+++ cid/cmd/main.go
@@ -3,10 +3,10 @@ package main
 import (
 	"os"
 
-	"github.com/starlinglab/integrity-v2/getcid"
+	"github.com/starlinglab/integrity-v2/cid"
 	"github.com/starlinglab/integrity-v2/util"
 )
 
 func main() {
-	util.Runner(os.Args[1:], getcid.Run)
+	util.Runner(os.Args[1:], cid.Run)
 }
