+++ export/cmd/main.go
@@ -3,10 +3,10 @@ package main
 import (
 	"os"
 
-	"github.com/starlinglab/integrity-v2/attr"
+	"github.com/starlinglab/integrity-v2/export"
 	"github.com/starlinglab/integrity-v2/util"
 )
 
 func main() {
-	util.Runner(os.Args[1:], attr.Run)
+	util.Runner(os.Args[1:], export.Run)
 }
